* master: (68 commits)
  DMPlex: More fixup from bad rebase, moved to DMGet/SetCoordinateSection()
  DM: Added DMGet/SetCoordinateSection()
  Fuckup: Fix for merge that is still untangling bad rebase from Oct. 17th 2013 - Fixed const in declarations - Fixed merge from PetscFE - Fixed 2 mallocs
  DMDA: Missing header - Damn rebase
  DMPlex ex3: Added 2D Q_1 and 3D P_2 tests
  DMDA: Added DMDAProjectFunction() and DMDAComputeL2Diff()
  DMDA: Fixed 2D geometry - Should reuse DMPlex routines
  DMDA: Fix bug in 2D closure operation for cells
  DMDA: Make closure operations also return the size
  DMDA: Changed DMDACreateSection() to emulate DMPlexCreateSection()
  DMDA: Added functions which emulate DMPlex functionality
  DMDA: Now DMDAGetNumCells() returns the cells in each direction as well
  PetscFE: Added support for tensor product cells in PetscDualSpace_Lagrange
  PetscFE: Added tensor product polynomial spaces
  example fix /sandbox/petsc/petsc.clone-3/include/petscmath.h(260): error: identifier "PETSC_CXX_STATIC_INLINE" is undefined   PETSC_STATIC_INLINE PetscReal PetscAbsScalar(PetscScalar a) {return a < 0.0 ? -a : a;}   ^
  remove warning in fun3d example "user.F", line 1256: warning: In-place macro substitution leaves line truncated "user.F", line 1259: warning: In-place macro substitution leaves line truncated "user.F", line 1262: warning: In-place macro substitution leaves line truncated
  /Users/petsc/petsc.clone-2/src/ksp/ksp/examples/tutorials/ex56.c:59: warning: comparison between signed and unsigned integer expressions
  fix examples for portability
  fixed to example outputs
  fix output for a few examples changed due to changes to -mat_view and not using -snes_monitor_short
  ...

Conflicts:
	src/dm/impls/plex/plexrefine.c

+++ src/ts/examples/tests/ex3.c
@@ -166,7 +166,7 @@ int main(int argc,char **argv)
   stepsz[0] = 1.0/(2.0*(nz-1)*(nz-1)); /* (mesh_size)^2/2.0 */
   ftime     = 0.0;
   for (k=0; k<nphase; k++) {
-    if (nphase > 1) printf("Phase %d: initial time %g, stepsz %g, duration: %g\n",k,ftime,stepsz[k],(k+1)*T);
+    if (nphase > 1) printf("Phase %d: initial time %g, stepsz %g, duration: %g\n",(int)k,ftime,stepsz[k],(k+1)*T);
     ierr = TSSetInitialTimeStep(ts,ftime,stepsz[k]);CHKERRQ(ierr);
     ierr = TSSetDuration(ts,max_steps,(k+1)*T);CHKERRQ(ierr);
